backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
mysqlclient>=2.2
django-cors-headers>=4.4
